GEODE-6886: Do not region sync if lost member is an empty accessor
  * Avoid region sync for empty accessor of persistent replicate region in correct place.
